handle for non w3c
