+++ lib/command/run-multiple.js
@@ -1,15 +1,13 @@
 const { fork } = require('child_process');
 const path = require('path');
 const crypto = require('crypto');
-const BStackLocal = require('browserstack-local').Local;
-const util = require('util');
 
 const runHook = require('../hooks');
 const event = require('../event');
 const collection = require('./run-multiple/collection');
 const { clearString, replaceValueDeep } = require('../utils');
 const {
-  getConfig, getTestRoot, fail,
+  getConfig, getTestRoot, fail, LocalTunnel,
 } = require('./utils');
 
 const runner = path.join(__dirname, '/../../bin/codecept');
@@ -27,57 +25,6 @@ let runId = 1;
 let subprocessCount = 0;
 let totalSubprocessCount = 0;
 let processesDone;
-
-class LocalTunnel {
-  constructor(config = {}) {
-    this.key = null;
-    this.browserstackLocal = false;
-    this.opts = {};
-    if (config.helpers && config.helpers.WebDriver && config.helpers.WebDriver.key) {
-      this.key = config.helpers.WebDriver.key;
-    }
-    if (config.plugins && config.plugins.wdio) {
-      if (config.plugins.wdio.key) {
-        this.key = config.plugins.wdio.key;
-      }
-      if (config.plugins.wdio.browserstackLocal) {
-        this.browserstackLocal = true;
-        config.plugins.wdio.browserstackLocal = false;
-      }
-      if (config.plugins.wdio.opts) {
-        this.opts = config.plugins.wdio.opts;
-      }
-    }
-    this.config = config;
-    this.local = new BStackLocal();
-  }
-
-  async start() {
-    if (this.browserstackLocal && this.key) {
-      try {
-        await util.promisify(this.local.start.bind(this.local))({ key: this.key, ...this.opts });
-        for (const key in this.config.multiple) {
-          if (this.config.multiple[key].browsers) {
-            this.config.multiple[key].browsers.forEach(browser => {
-              browser.desiredCapabilities && (browser.desiredCapabilities['bstack:options'].local = true);
-            });
-          }
-        }
-        this.localStarted = true;
-      } catch (err) {
-        console.log(err);
-      }
-    }
-  }
-
-  async stop() {
-    try {
-      this.localStarted && await util.promisify(this.local.stop.bind(this.local))();
-    } catch (err) {
-      console.log(err);
-    }
-  }
-}
 
 module.exports = async function (selectedRuns, options) {
   // registering options globally to use in config
