+++ lib/command/utils.js
@@ -3,6 +3,7 @@ const path = require('path');
 const semver = require('semver');
 const util = require('util');
 const mkdirp = require('mkdirp');
+const BStackLocal = require('browserstack-local').Local;
 
 const output = require('../output');
 const { fileExists, beautify } = require('../utils');
@@ -112,5 +113,62 @@ module.exports.createOutputDir = (config, testRoot) => {
   if (!fileExists(outputDir)) {
     output.print(`creating output directory: ${outputDir}`);
     mkdirp.sync(outputDir);
+  }
+};
+
+module.exports.LocalTunnel = class LocalTunnel {
+  constructor(config = {}) {
+    this.key = null;
+    this.browserstackLocal = false;
+    this.opts = {};
+    if (config.helpers && config.helpers.WebDriver && config.helpers.WebDriver.key) {
+      this.key = config.helpers.WebDriver.key;
+    }
+    if (config.plugins && config.plugins.wdio) {
+      if (config.plugins.wdio.key) {
+        this.key = config.plugins.wdio.key;
+      }
+      if (config.plugins.wdio.browserstackLocal) {
+        this.browserstackLocal = true;
+        config.plugins.wdio.browserstackLocal = false;
+      }
+      if (config.plugins.wdio.opts) {
+        this.opts = config.plugins.wdio.opts;
+      }
+    }
+    this.config = config;
+    this.local = new BStackLocal();
+  }
+
+  async start() {
+    if (this.browserstackLocal && this.key) {
+      try {
+        await util.promisify(this.local.start.bind(this.local))({ key: this.key, ...this.opts });
+        for (const key in this.config.multiple) {
+          if (this.config.multiple[key].browsers) {
+            this.config.multiple[key].browsers.forEach(browser => {
+              if (browser.desiredCapabilities) {
+                if (browser.desiredCapabilities['bstack:options']) {
+                  browser.desiredCapabilities['bstack:options'].local = true;
+                } else {
+                  browser.desiredCapabilities['browserstack.local'] = true;
+                }
+              }
+            });
+          }
+        }
+        this.localStarted = true;
+      } catch (err) {
+        console.log(err);
+      }
+    }
+  }
+
+  async stop() {
+    try {
+      this.localStarted && await util.promisify(this.local.stop.bind(this.local))();
+    } catch (err) {
+      console.log(err);
+    }
   }
 };
